xythrion/pyproject.toml
[project]
name = "Xythrion"
version = "0.1.0"
description = "Discord graphing bot"
authors = [
    {name = "Xithrius", email = "xithrius@gmail.com"},
]
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "discord-py>=2.3.2",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.24.1",
    "pillow>=10.0.0",
    "croniter>=1.4.1",
]

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
lint = [
    "ruff>=0.0.285"
]
format = [
    "black>=23.7.0"
]
precommit = [
    "pre-commit>=3.3.3",
]
test = [
    "pytest>=7.4.0",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.scripts]
start = "python -m bot"
lint = "pre-commit run --all-files"
precommit = "pre-commit install"
test = "pytest"

[tool.ruff]
target-version = "py311"
extend-exclude = [".cache"]
ignore = [
    "ANN002", "ANN003", "ANN101", "ANN102", "ANN204", "ANN206", "ANN401",
    "B904",
    "C401", "C408",
    "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D203", "D212", "D214", "D215", "D301",
    "D400", "D401", "D402", "D404", "D405", "D406", "D407", "D408", "D409", "D410", "D411", "D412", "D413", "D414", "D416", "D417",
    "E731",
    "RET504",
    "RUF005",
    "S101", "S311",
    "SIM102", "SIM108",
]
line-length = 120
select = ["ANN", "B", "C4", "D", "DTZ", "E", "F", "ISC", "N", "PGH", "PIE", "Q", "RET", "RSE", "RUF", "S", "SIM", "T20", "TID", "UP", "W"]

# [tool.ruff.per-file-ignores]
# "/tests/*" = ["ANN", "D"]
